Make pytest ignore conf.py
The doctest-modules option causes it to error on conf.py unless
sphinx_bootstrap_theme is installed. I don't want to add that as a dev
requirement because most people won't need it for dev or testing.
There's no reason pytest needs to look at conf.py, so we can just
exclude it.

In the future I could put all of these pytest options into a pytest.ini
file, but I'm going to wait on that for now.