Fix bugs in merge_columns_by_label and make it ignore None/np.nan
For monotonically indexed dataframes, get_loc() returns a slice instead
of a mask.  I changed the way the first_iloc is obtained so that it will
work with all dataframes. Also there are some cases where try to change
a column value to an iterable will result in unexpected exceptions. I
fixed this by adding a new column in the position of the original
instead of changing the original.